Fix error when TLS 1.3 is not available.
In version 1.0.1 and older, the module may fail to connect to NinjaRMM
if the underlying operating system and/or PowerShell host did not
support TLS 1.3.

This has been corrected by enabling TLS 1.2 (which is available on all
systems that support PowerShell 5+), then performing a check to see if
TLS 1.3 exists and can be enabled.